refactor(project-structure): organize files into domain, controller, view, util packages
