+++ src/main/kotlin/lotto/Application.kt
@@ -1,5 +1,7 @@
 package lotto
 
+import lotto.controller.LottoManager
+
 fun main() {
     LottoManager.run()
 }
