+++ src/main/kotlin/lotto/domain/LottoMachine.kt
@@ -1,6 +1,7 @@
-package lotto
+package lotto.domain
 
 import camp.nextstep.edu.missionutils.Randoms
+import lotto.Lotto
 
 object LottoMachine {
     fun issueTickets(ticketCount: Int): List<Lotto> {
