+++ src/main/kotlin/lotto/domain/LottoRank.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 enum class LottoRank(
     val matchCount: Int,
