+++ src/main/kotlin/lotto/domain/LottoResult.kt
@@ -1,4 +1,6 @@
-package lotto
+package lotto.domain
+
+import lotto.Lotto
 
 class LottoResult(
