+++ src/main/kotlin/lotto/util/InputParser.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.util
 
 object InputParser {
