+++ src/main/kotlin/lotto/util/InputValidator.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.util
 
 object InputValidator {
