+++ src/test/kotlin/lotto/InputValidatorTest.kt
@@ -1,5 +1,6 @@
 package lotto
 
+import lotto.util.InputValidator
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.api.Test
 
@@ -7,7 +8,7 @@ class InputValidatorTest {
 
     @Test
     fun `should throw exception when purchase amount is not a multiple of 1000`() {
-        assertThatThrownBy {InputValidator.validatePurchaseAmount(1500) }
+        assertThatThrownBy { InputValidator.validatePurchaseAmount(1500) }
             .isInstanceOf(IllegalArgumentException::class.java)
             .hasMessage("[ERROR] The purchase amount must be divisible by 1,000")
     }
